Fix stupid bug.

+++ data/data_generation/tools.py
@@ -192,11 +192,11 @@ def get_all_cf(pos_gal_fns, pos_mat_fns, Lbox):
 
     # obtain all power spectra
     cf_obj = SimulationBox2PCF(mode='1d', data1=cat_gal, data2=cat_gal, edges=r, BoxSize=Lbox)
-    corr = cf_obj.corr.astype(np.float64)
-    D1D2 = cf_obj.D1D2.astype(np.float64)
+    corr = cf_obj.corr
+    D1D2 = cf_obj.D1D2
     # D1R2 = cf_obj.D1R2.astype(np.float64)
     # D2R1 = cf_obj.D2R1.astype(np.float64)
-    R1R2 = cf_obj.R1R2.astype(np.float64)
+    R1R2 = cf_obj.R1R2
     cf_dic['corr_gg'] = corr['corr']
     cf_dic['D1D2_gg'] = D1D2['npairs']
     # cf_dic['D1R2_gg'] = D1R2
@@ -208,11 +208,11 @@ def get_all_cf(pos_gal_fns, pos_mat_fns, Lbox):
     # create catalog from fitsfile
     cat_mat = FITSCatalog(pos_mat_fns, ext='Data')
     cf_obj = SimulationBox2PCF(mode='1d', data1=cat_gal, data2=cat_mat, edges=r, BoxSize=Lbox)
-    corr = cf_obj.corr.astype(np.float64)
-    D1D2 = cf_obj.D1D2.astype(np.float64)
+    corr = cf_obj.corr
+    D1D2 = cf_obj.D1D2
     # D1R2 = cf_obj.D1R2.astype(np.float64)
     # D2R1 = cf_obj.D2R1.astype(np.float64)
-    R1R2 = cf_obj.R1R2.astype(np.float64)
+    R1R2 = cf_obj.R1R2
     cf_dic['corr_gm'] = corr['corr']
     cf_dic['D1D2_gm'] = D1D2['npairs']
     # cf_dic['D1R2_gm'] = D1R2
